Use fake user to emulate login/logout
