+++ client/.env.local
@@ -1,1 +1,2 @@
 API_BASE_PATH=http://localhost:8080
+USE_FAKE_USER=true
