+++ client/src/app/Header.tsx
@@ -1,59 +1,16 @@
-import { Alignment, Navbar, Classes, Button, Intent, Popover, Menu, Dialog } from '@blueprintjs/core';
-import { IconNames } from '@blueprintjs/icons';
-import React, { useContext, useEffect, useState } from 'react';
-import { Link } from 'react-router-dom';
+import {Alignment, Classes, Menu, Navbar, Popover} from '@blueprintjs/core';
+import {IconNames} from '@blueprintjs/icons';
+import React, {useContext} from 'react';
+import {Link} from 'react-router-dom';
 import cn from 'classnames';
-import LoggedUserContext from '../services/loggedUserContext';
-import { LoginProvider } from '../model';
-import ServicesContext from '../services/servicesContext';
+import LoggedUserContext from '../login/loggedUserContext';
+import {LoginButton} from '../login/LoginButton';
 
 export const AddStoryButton: React.FC = () => (
   <Link className={cn(Classes.BUTTON, Classes.INTENT_PRIMARY)} to="/add-story">
     Добавить свою историю
   </Link>
 );
-
-function loginViaProvider(provider: LoginProvider) {
-  const loginUrl = new URL(provider.auth_url);
-  loginUrl.searchParams.set('client_id', provider.client_id);
-  loginUrl.searchParams.set('redirect_uri', provider.redirect_url);
-
-  if (provider.scope) {
-    loginUrl.searchParams.set('scope', provider.scope);
-  }
-
-  window.location.assign(loginUrl.href);
-}
-
-const LoginButton: React.FC = () => {
-  const [dialogIsOpen, setDialogIsOpen] = useState(false);
-  const [loginProviders, setLoginProviders] = useState<LoginProvider[]>([]);
-  const services = useContext(ServicesContext);
-
-  useEffect(() => {
-    if (dialogIsOpen && loginProviders.length === 0) {
-      services.apiClient.getLoginProviders().then(setLoginProviders);
-    }
-  }, [dialogIsOpen, loginProviders, services]);
-
-  return (
-    <>
-      <Button intent={Intent.PRIMARY} onClick={() => setDialogIsOpen(true)}>
-        Войти
-      </Button>
-      <Dialog isOpen={dialogIsOpen} onClose={() => setDialogIsOpen(false)} title="Вход">
-        <div className={Classes.DIALOG_BODY}>
-          <p>Войдите через одну из социальных сетей, чтобы рассказать свою историю</p>
-          {loginProviders.map(provider => (
-            <div key={provider.name}>
-              <Button onClick={() => loginViaProvider(provider)}>{provider.name}</Button>
-            </div>
-          ))}
-        </div>
-      </Dialog>
-    </>
-  );
-};
 
 const Header: React.FC = ({ children }) => {
   const { loggedUser, logout } = useContext(LoggedUserContext);
@@ -74,7 +31,7 @@ const Header: React.FC = ({ children }) => {
           </Menu>
         }
       >
-        <img src={loggedUser.imageUrl} alt={loggedUser.name} className="avatar" />
+        <img src={loggedUser.ImageURL} alt={loggedUser.Username} className="avatar" />
       </Popover>
     </>
   ) : (
