+++ client/src/modules.d.ts
@@ -9,7 +9,8 @@ declare const module: Module;
 type Process = {
   env: {
     API_BASE_PATH: string;
+    USE_FAKE_USER: boolean;
   };
 };
 
-declare const process: Process
+declare const process: Process;
